fix(typescript): export all the types
